feat(session-ui): implement lazy loading for trace rows
