+++ web/src/components/session/index.tsx
@@ -5,7 +5,6 @@ import { StarSessionToggle } from "@/src/components/star-toggle";
 import { IOPreview } from "@/src/components/trace/IOPreview";
 import { JsonSkeleton } from "@/src/components/ui/CodeJsonViewer";
 import { Badge } from "@/src/components/ui/badge";
-import { Card } from "@/src/components/ui/card";
 import { DetailPageNav } from "@/src/features/navigate-detail-pages/DetailPageNav";
 import { useDetailPageLists } from "@/src/features/navigate-detail-pages/context";
 import { api } from "@/src/utils/api";
@@ -34,184 +33,12 @@ import { CreateNewAnnotationQueueItem } from "@/src/features/annotation-queues/c
 import { TablePeekView } from "@/src/components/table/peek";
 import { PeekViewTraceDetail } from "@/src/components/table/peek/peek-trace-detail";
 import { usePeekNavigation } from "@/src/components/table/peek/hooks/usePeekNavigation";
-import { NewDatasetItemFromExistingObject } from "@/src/features/datasets/components/NewDatasetItemFromExistingObject";
-import { ItemBadge } from "@/src/components/ItemBadge";
 import { type WithStringifiedMetadata } from "@/src/utils/clientSideDomainTypes";
+import { LazyTraceRow } from "@/src/components/session/TraceRow";
 
 // some projects have thousands of users in a session, paginate to avoid rendering all at once
 const INITIAL_USERS_DISPLAY_COUNT = 10;
 const USERS_PER_PAGE_IN_POPOVER = 50;
-
-// Skeleton placeholder for trace cards
-const TraceSkeleton = () => {
-  return (
-    <Card className="border-border shadow-none">
-      <div className="flex h-64 items-center justify-center p-4">
-        <JsonSkeleton className="h-full w-full" numRows={8} />
-      </div>
-    </Card>
-  );
-};
-
-// Real trace card with all the heavy content (memoized to prevent unnecessary re-renders)
-const TraceCard = React.memo(
-  ({
-    trace,
-    projectId,
-    openPeek,
-    traceCommentCounts,
-  }: {
-    trace: any;
-    projectId: string;
-    openPeek: (id: string, row: any) => void;
-    traceCommentCounts: any;
-  }) => {
-    return (
-      <Card className="border-border shadow-none">
-        <div className="grid md:grid-cols-[1fr_1px_358px] lg:grid-cols-[1fr_1px_28rem]">
-          <div className="overflow-hidden py-4 pl-4 pr-4">
-            <SessionIO
-              traceId={trace.id}
-              projectId={projectId}
-              timestamp={new Date(trace.timestamp)}
-            />
-          </div>
-          <div className="hidden bg-border md:block"></div>
-          <div className="flex flex-col border-t py-4 pl-4 pr-4 md:border-0">
-            <div className="mb-4 flex flex-col gap-2">
-              <Link
-                href={`/project/${projectId}/traces/${trace.id}`}
-                className="flex items-start gap-2 rounded-lg border p-2 transition-colors hover:bg-accent"
-                onClick={(e) => {
-                  if (!e.metaKey && !e.ctrlKey && !e.shiftKey) {
-                    e.preventDefault();
-                    openPeek(trace.id, trace);
-                  }
-                }}
-              >
-                <ItemBadge type="TRACE" isSmall />
-                <div className="flex flex-col">
-                  <span className="text-xs font-medium">
-                    {trace.name} ({trace.id})&nbsp;↗
-                  </span>
-                  <span className="text-xs text-muted-foreground">
-                    {trace.timestamp.toLocaleString()}
-                  </span>
-                </div>
-              </Link>
-              <div className="flex flex-wrap gap-2">
-                <NewDatasetItemFromTraceId
-                  projectId={projectId}
-                  traceId={trace.id}
-                  timestamp={new Date(trace.timestamp)}
-                  buttonVariant="outline"
-                />
-                <AnnotateDrawer
-                  key={"annotation-drawer" + trace.id}
-                  projectId={projectId}
-                  scoreTarget={{
-                    type: "trace",
-                    traceId: trace.id,
-                  }}
-                  scores={trace.scores}
-                  buttonVariant="outline"
-                  analyticsData={{
-                    type: "trace",
-                    source: "SessionDetail",
-                  }}
-                  scoreMetadata={{
-                    projectId: projectId,
-                    environment: trace.environment,
-                  }}
-                />
-                <CommentDrawerButton
-                  projectId={projectId}
-                  variant="outline"
-                  objectId={trace.id}
-                  objectType="TRACE"
-                  count={getNumberFromMap(traceCommentCounts, trace.id)}
-                />
-              </div>
-            </div>
-            <div className="flex-1">
-              <p className="mb-1 font-medium">Scores</p>
-              <div className="flex flex-wrap content-start items-start gap-1">
-                <GroupedScoreBadges scores={trace.scores} />
-              </div>
-            </div>
-          </div>
-        </div>
-      </Card>
-    );
-  },
-);
-TraceCard.displayName = "TraceCard";
-
-// Hydrating component: Starts as skeleton, becomes real when visible
-const HydratingTraceCard = React.forwardRef<
-  HTMLDivElement,
-  {
-    trace: any;
-    projectId: string;
-    openPeek: (id: string, row: any) => void;
-    traceCommentCounts: any;
-    index: number;
-    onHydrated?: (index: number) => void;
-  }
->((props, measureRef) => {
-  const { index, onHydrated, ...cardProps } = props;
-  const [shouldHydrate, setShouldHydrate] = useState(false);
-  const intersectionRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const node = intersectionRef.current;
-    if (!node) return;
-
-    const observer = new IntersectionObserver(
-      ([entry]) => {
-        if (entry?.isIntersecting) {
-          setShouldHydrate(true);
-          observer.disconnect(); // Only hydrate once
-        }
-      },
-      { rootMargin: "400px" }, // Pre-load content 400px before visible
-    );
-
-    observer.observe(node);
-    return () => observer.disconnect();
-  }, []);
-
-  // Notify parent when hydration completes so it can remeasure
-  useEffect(() => {
-    if (shouldHydrate && onHydrated) {
-      // Use setTimeout to ensure DOM has updated
-      setTimeout(() => {
-        onHydrated(index);
-      }, 0);
-    }
-  }, [shouldHydrate, onHydrated, index]);
-
-  // Combine refs: virtualizer measures, IO observes
-  const combinedRef = useCallback(
-    (node: HTMLDivElement | null) => {
-      if (typeof measureRef === "function") {
-        measureRef(node);
-      } else if (measureRef) {
-        (measureRef as React.MutableRefObject<HTMLDivElement | null>).current =
-          node;
-      }
-      intersectionRef.current = node;
-    },
-    [measureRef],
-  );
-
-  return (
-    <div ref={combinedRef} className="pb-3">
-      {shouldHydrate ? <TraceCard {...cardProps} /> : <TraceSkeleton />}
-    </div>
-  );
-});
-HydratingTraceCard.displayName = "HydratingTraceCard";
 
 export function SessionUsers({
   projectId,
@@ -578,14 +405,14 @@ export const SessionPage: React.FC<{
                     transform: `translateY(${virtualItem.start}px)`,
                   }}
                 >
-                  <HydratingTraceCard
+                  <LazyTraceRow
                     ref={virtualizer.measureElement}
                     trace={trace}
                     projectId={projectId}
                     openPeek={openPeek}
                     traceCommentCounts={traceCommentCounts.data}
                     index={virtualItem.index}
-                    onHydrated={() => {
+                    onLoad={() => {
                       // Force virtualizer to remeasure this specific item
                       virtualizer.measureElement(
                         document.querySelector(
@@ -657,43 +484,5 @@ export const SessionIO = ({
         </div>
       )}
     </div>
-  );
-};
-
-const NewDatasetItemFromTraceId = (props: {
-  projectId: string;
-  traceId: string;
-  timestamp: Date;
-  buttonVariant?: "outline" | "secondary";
-}) => {
-  // SessionIO already fetches the trace, so this doesn't add an extra request
-  const trace = api.traces.byId.useQuery(
-    {
-      traceId: props.traceId,
-      projectId: props.projectId,
-      timestamp: props.timestamp,
-    },
-    {
-      enabled: typeof props.traceId === "string",
-      trpc: {
-        context: {
-          skipBatch: true,
-        },
-      },
-      refetchOnMount: false,
-    },
-  );
-
-  if (!trace.data) return null;
-
-  return (
-    <NewDatasetItemFromExistingObject
-      projectId={props.projectId}
-      traceId={props.traceId}
-      input={trace.data.input ?? null}
-      output={trace.data.output ?? null}
-      metadata={trace.data.metadata ?? null}
-      buttonVariant={props.buttonVariant}
-    />
   );
 };
